+++ src/app/modules/appsConfigurations/domain.ts
@@ -1,18 +1,20 @@
 import { IdWithIdentifier as JsonIdWithIdentifier } from '../../../YamlSettings.interface';
 
+import { Rect } from '../shared/app/Rect';
+
 export enum ApplicationOptions {
   Float = 'float',
   Unmanage = 'unmanage',
   Force = 'force',
   ObjectNameChange = 'object_name_change',
   Layered = 'layered',
-  BorderOverflow = 'border_overflow',
+  /*BorderOverflow = 'border_overflow',*/ // present in komorebi but deprecated
   TrayAndMultiWindow = 'tray_and_multi_window',
 }
 
 export const LabelByAppOption: Record<ApplicationOptions, string> = {
   [ApplicationOptions.Float]: 'Float',
-  [ApplicationOptions.BorderOverflow]: 'Border Overflow',
+  /*[ApplicationOptions.BorderOverflow]: 'Border Overflow',*/
   [ApplicationOptions.Force]: 'Forced',
   [ApplicationOptions.Layered]: 'Layered',
   [ApplicationOptions.ObjectNameChange]: 'Name Change',
@@ -50,6 +52,7 @@ export interface AppConfiguration extends AppConfigurationsOptions {
   kind: ApplicationIdentifier;
   identifier: string;
   matchingStrategy: MatchingStrategy;
+  invisibleBorders: Rect.plain | null;
 }
 
 export class AppConfiguration {
@@ -63,12 +66,13 @@ export class AppConfiguration {
       identifier: 'new-app.exe',
       matchingStrategy: MatchingStrategy.Equals,
       [ApplicationOptions.Float]: false,
-      [ApplicationOptions.BorderOverflow]: false,
+      /*[ApplicationOptions.BorderOverflow]: false,*/
       [ApplicationOptions.Force]: false,
       [ApplicationOptions.Layered]: false,
       [ApplicationOptions.ObjectNameChange]: false,
       [ApplicationOptions.TrayAndMultiWindow]: false,
       [ApplicationOptions.Unmanage]: false,
+      invisibleBorders: null,
     };
   }