+++ src/app/modules/appsConfigurations/infra/EditModal.tsx
@@ -1,8 +1,9 @@
 import { SettingsGroup, SettingsOption, SettingsSubGroup } from '../../../components/SettingsBox';
-import { Input, Modal, Select, Switch } from 'antd';
+import { Button, Input, InputNumber, Modal, Select, Switch } from 'antd';
 import React, { useState } from 'react';
 
 import { useAppSelector } from '../../shared/app/hooks';
+import { Rect } from '../../shared/app/Rect';
 import { RootSelectors } from '../../shared/app/selectors';
 import { OptionsFromEnum } from '../../shared/app/utils';
 
@@ -19,12 +20,12 @@ interface Props {
 }
 
 export const EditAppModal = ({ idx, onCancel, onSave, isNew, open }: Props) => {
+  const monitors = useAppSelector(RootSelectors.monitors);
   const _app = useAppSelector((state) => {
     return idx != null && !isNew ? state.appsConfigurations[idx]! : AppConfiguration.default();
   })!;
   const [app, setApp] = useState(_app);
-
-  const monitors = useAppSelector(RootSelectors.monitors);
+  const { invisibleBorders } = app;
 
   const onInternalSave = () => {
     onSave(app);
@@ -43,10 +44,22 @@ export const EditAppModal = ({ idx, onCancel, onSave, isNew, open }: Props) => {
 
   const onChangeOption = (option: ApplicationOptions, value: boolean) => setApp({ ...app, [option]: value });
 
+  const resetInvisibleBorders = () => setApp({ ...app, invisibleBorders: null });
+  const onChangeInvisibleBorders = (side: keyof Rect.plain, value: number | null) => {
+    setApp({
+      ...app,
+      invisibleBorders: {
+        ...(app.invisibleBorders || new Rect().plain()),
+        [side]: value || 0,
+      },
+    });
+  };
+
   const monitorsOptions = monitors.map((_, i) => ({ label: `Monitor ${i + 1}`, value: i }));
-  const workspaceOptions = app.monitor != null && monitors[app.monitor]
-    ? monitors[app.monitor]?.workspaces.map(({ name }) => ({ label: name, value: name }))
-    : [];
+  const workspaceOptions =
+    app.monitor != null && monitors[app.monitor]
+      ? monitors[app.monitor]?.workspaces.map(({ name }) => ({ label: name, value: name }))
+      : [];
 
   return (
     <Modal
@@ -122,6 +135,48 @@ export const EditAppModal = ({ idx, onCancel, onSave, isNew, open }: Props) => {
               <Switch value={app[value]} onChange={onChangeOption.bind(this, value)} />
             </SettingsOption>
           ))}
+        </SettingsSubGroup>
+      </SettingsGroup>
+
+      <SettingsGroup>
+        <SettingsSubGroup
+          label={
+            <SettingsOption>
+              <span>Specifit invisible borders</span>
+              <Button type="dashed" onClick={resetInvisibleBorders}>
+                ⟳
+              </Button>
+            </SettingsOption>
+          }
+        >
+          <SettingsOption>
+            <span>Left</span>
+            <InputNumber
+              value={invisibleBorders?.left}
+              onChange={onChangeInvisibleBorders.bind(this, 'left')}
+              placeholder="Global"
+            />
+          </SettingsOption>
+          <SettingsOption>
+            <span>Top</span>
+            <InputNumber value={invisibleBorders?.top} onChange={onChangeInvisibleBorders.bind(this, 'top')} placeholder="Global" />
+          </SettingsOption>
+          <SettingsOption>
+            <span>Right</span>
+            <InputNumber
+              value={invisibleBorders?.right}
+              onChange={onChangeInvisibleBorders.bind(this, 'right')}
+              placeholder="Global"
+            />
+          </SettingsOption>
+          <SettingsOption>
+            <span>Bottom</span>
+            <InputNumber
+              value={invisibleBorders?.bottom}
+              onChange={onChangeInvisibleBorders.bind(this, 'bottom')}
+              placeholder="Global"
+            />
+          </SettingsOption>
         </SettingsSubGroup>
       </SettingsGroup>
     </Modal>