+++ changelog.md
@@ -1,15 +1,22 @@
 # Changelog
 
+## [Unreleased]
+### added
+- new feature of invisible borders per app
+
+### changed
+- delete border overflow and changed for invisible borders per app
+
 ## [1.0.0-beta.3]
 ### added
 - new apps templates
 - add AHK as a dependency to show to new users
 - add AHK settings
 
-
 ## [1.0.0-beta.2]
 ### added
 - export option for apps
+
 ### fixed
 - delete binded monitor and workpace on an application
 - bad intallation on setup